+++ src/pages/register/success/index.tsx
@@ -12,7 +12,7 @@ export default function RegisterSuccess() {
   const [isLoaded, setIsLoaded] = useState<boolean>(false);
   const goToHome = () => {
     setIsLoaded(true);
-    router.push('/').catch((error) => {
+    router.push('/home').catch((error) => {
       console.error('페이지 이동에 실패하였습니다.', error);
     });
   };
